refactor(task): workflowモジュールをarticleにリネーム
既存のworkflowモジュールをより具体的なarticleに名称変更し、関連するテストも移動
